Update version to 2.0.2 and register new network packets
This commit updates the version of the mod from 2.0.1 to 2.0.2 in both the mod.json and build.gradle files. In addition, the registration of network packets has been centralized in the main class "Contracts.java". This makes the code cleaner and avoids duplicate registration in other classes such as "ServerActionType.java" and "ClientActionType.java".